+++ src/app/mork/celery/tasks/deletion.py
@@ -9,6 +9,7 @@ from sqlalchemy import insert, select
 from sqlalchemy.exc import SQLAlchemyError
 
 from mork.celery.celery_app import app
+from mork.celery.tasks.brevo import delete_brevo_platform_user
 from mork.celery.tasks.edx import delete_edx_platform_user
 from mork.conf import settings
 from mork.db import MorkDB
@@ -76,7 +77,7 @@ def delete_user(
     delete_chain = chain(
         remove_email_status.si(email),
         mark_user_for_deletion.si(email=email, reason=reason),
-        delete_edx_platform_user.s(),
+        group(delete_edx_platform_user.s(), delete_brevo_platform_user.s()),
     )
 
     delete_chain.delay()